+++ src/main/java/com/example/sklad/model/dto/documents/OrderDTO.java
@@ -38,8 +38,8 @@ public class OrderDTO {
     @JsonProperty("payment_type")
     private String paymentType;
 
-    private double amount;
-    private double tax;
+    private float amount;
+    private float tax;
 
     @JsonProperty("is_hold")
     private boolean isHold;
