+++ src/main/java/com/example/sklad/repositories/CompanyRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Long> {
+public interface CompanyRepository extends JpaRepository<Company, Integer> {
     Optional<Company> findByNameIgnoreCase(String name);
 }
