+++ src/main/java/com/example/sklad/repositories/PriceRepository.java
@@ -3,5 +3,5 @@ package com.example.sklad.repositories;
 import com.example.sklad.model.entities.Price;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PriceRepository extends JpaRepository<Price, Long> {
+public interface PriceRepository extends JpaRepository<Price, Integer> {
 }
