+++ src/main/java/com/example/sklad/repositories/StorageRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StorageRepository extends JpaRepository<Storage, Long> {
+public interface StorageRepository extends JpaRepository<Storage, Integer> {
 
     Optional<Storage> findByNameIgnoreCase(String name);
